created a library with all main files that can now be used from anywhere, all test methods still pass and are working correctly
