+++ CsvFileConverter.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CsvFileConverter", "InternP
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CsvFileConverterTests", "InternProject.CsvFileConverter.Tests\CsvFileConverterTests.csproj", "{1D480948-73E2-4C6B-8DBC-32A6741864CB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InternProject.CsvFileConverter.Library", "InternProject.CsvFileConverter.Library\InternProject.CsvFileConverter.Library.csproj", "{621F7B76-163E-4C4B-98A1-D58E7AC8B374}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{1D480948-73E2-4C6B-8DBC-32A6741864CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1D480948-73E2-4C6B-8DBC-32A6741864CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1D480948-73E2-4C6B-8DBC-32A6741864CB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{621F7B76-163E-4C4B-98A1-D58E7AC8B374}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{621F7B76-163E-4C4B-98A1-D58E7AC8B374}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{621F7B76-163E-4C4B-98A1-D58E7AC8B374}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{621F7B76-163E-4C4B-98A1-D58E7AC8B374}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
