+++ InternProject.CsvFileConverter.Library/Extensions/Formatters/FormatterType.cs
@@ -1,0 +1,8 @@
+﻿namespace InternProject.CsvFileConverter.Library
+{
+    public enum FormatterType
+    {
+        Json,
+        Xml
+    }
+}
