+++ InternProject.CsvFileConverter.Library/Extensions/Formatters/JsonTextFormatter.cs
@@ -1,0 +1,14 @@
+﻿using Newtonsoft.Json;
+
+namespace InternProject.CsvFileConverter.Library
+{
+    public class JsonTextFormatter : ITextFormatter
+    {
+        public FormatterType Type => FormatterType.Json;
+
+        public string Format<T>(T t) where T : class
+        {
+            return JsonConvert.SerializeObject(t);
+        }
+    }
+}
