+++ InternProject.CsvFileConverter.Library/Interfaces/Core.Conversions.Interfaces/IDataExtractor.cs
@@ -1,0 +1,9 @@
+﻿using System.IO;
+
+namespace InternProject.CsvFileConverter.Library
+{
+    public interface IDataExtractor
+    {
+        DealData[] ReadContent(StringReader input, bool headerPresent);
+    }
+}
