+++ InternProject.CsvFileConverter.Library/Interfaces/Core.Conversions.Interfaces/IJsonConverter.cs
@@ -1,0 +1,7 @@
+﻿namespace InternProject.CsvFileConverter.Library
+{
+    public interface IJsonConverter
+    {
+        string ConvertToJson(DealData[] data);
+    }
+}
