+++ InternProject.CsvFileConverter.Library/Interfaces/Core.IO.Interfaces/IFileReader.cs
@@ -1,0 +1,9 @@
+﻿using System.IO;
+
+namespace InternProject.CsvFileConverter.Library
+{
+    public interface IFileReader
+    {
+        StringReader ReadContent(string input);
+    }
+}
