+++ InternProject.CsvFileConverter.Library/Interfaces/Extensions.Interfaces/ITextFormatter.cs
@@ -1,0 +1,9 @@
+﻿namespace InternProject.CsvFileConverter.Library
+{
+    public interface ITextFormatter
+    {
+        FormatterType Type { get; }
+
+        string Format<T>(T t) where T : class;
+    }
+}
