+++ InternProject.CsvFileConverter.Library/Interfaces/Validation.Interface/IFieldValidator.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace InternProject.CsvFileConverter.Library
+{
+    public interface IFieldValidator
+    {
+        Type TypeToValidate { get; }
+
+        FieldValidationResult Validate(string fieldValue);
+    }
+}
