+++ InternProject.CsvFileConverter.Library/Validations/StringFieldValidator.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace InternProject.CsvFileConverter.Library
+{
+    public class StringFieldValidator : IFieldValidator
+    {
+        public Type TypeToValidate => typeof(string);
+
+        public FieldValidationResult Validate(string fieldValue)
+        {
+            return new FieldValidationResult(fieldValue);
+        }
+    }
+}
