+++ InternProject.CsvFileConverter/CsvFileConverter.csproj
@@ -14,6 +14,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Files\**" />
+    <EmbeddedResource Remove="Files\**" />
+    <None Remove="Files\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Autofac" Version="4.8.0" />
     <PackageReference Include="CsvHelper" Version="7.1.0" />
     <PackageReference Include="FluentAssertions" Version="5.3.0" />
@@ -29,7 +35,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Files\" />
+    <ProjectReference Include="..\InternProject.CsvFileConverter.Library\InternProject.CsvFileConverter.Library.csproj" />
   </ItemGroup>
 
 </Project>
