--- InternProject.CsvFileConverter/Extensions/Formatters/FormatterType.cs
@@ -1,8 +1,0 @@
-﻿namespace CsvFileConverter
-{
-    public enum FormatterType
-    {
-        Json,
-        Xml
-    }
-}
