--- InternProject.CsvFileConverter/Interfaces/Core.Conversions.Interfaces/IDataExtractor.cs
@@ -1,11 +1,0 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using System.IO;
-
-namespace CsvFileConverter
-{
-    public interface IDataExtractor
-    {
-        DealData[] ReadContent(StringReader input, bool headerPresent);
-       
-    }
-}
