--- InternProject.CsvFileConverter/Interfaces/Core.Conversions.Interfaces/IJsonConverter.cs
@@ -1,7 +1,0 @@
-﻿namespace CsvFileConverter
-{
-    public interface IJsonConverter
-    {
-        string ConvertToJson(DealData[] data);
-    }
-}
