--- InternProject.CsvFileConverter/Interfaces/Core.IO.Interfaces/IFileReader.cs
@@ -1,9 +1,0 @@
-﻿using System.IO;
-
-namespace CsvFileConverter.MainProgramme
-{
-    public interface IFileReader
-    {
-        StringReader ReadContent(string input);
-    }
-}
