--- InternProject.CsvFileConverter/Interfaces/Core.IO.Interfaces/IFileWriter.cs
@@ -1,10 +1,0 @@
-﻿namespace CsvFileConverter
-{
-    public interface IFileWriter
-    {
-        void WriteContent(string output, DealData[] data);
-        void WriteContent(string output, DealData[] data, bool overwrite);
-        void WriteContent(string output, DealData[] data, FormatterType formatterType);
-        void WriteContent(string output, DealData[] data, bool overwrite, FormatterType formatterType);
-    }
-}
