--- InternProject.CsvFileConverter/Interfaces/Extensions.Interfaces/ITextFormatter.cs
@@ -1,9 +1,0 @@
-﻿namespace CsvFileConverter
-{
-    public interface ITextFormatter
-    {
-        FormatterType Type { get; }
-
-        string Format<T>(T t) where T : class;
-    }
-}
