--- InternProject.CsvFileConverter/Interfaces/Validation.Interface/IFieldValidator.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace CsvFileConverter
-{
-    public interface IFieldValidator
-    {
-        Type TypeToValidate { get; }
-
-        FieldValidationResult Validate(string fieldValue);
-    }
-}
