--- InternProject.CsvFileConverter/Logging/LoggerConfigFile.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Text;
-using Serilog;
-using Serilog.Exceptions;
-
-namespace CsvFileConverter.Logging
-{
-    [ExcludeFromCodeCoverage]
-    class LoggerConfigFile 
-    {
-
-        public ILogger SeriLogConfig => Log.Logger = new LoggerConfiguration()
-                .Enrich.FromLogContext()
-                .Enrich.WithExceptionDetails()
-                .Enrich.WithCaller()
-                .WriteTo.File("C:\\GIT\\InternProject\\Stage3_Verification\\Logging\\CsvToJson.log",
-                    outputTemplate:
-                    "{Timestamp:HH:mm:ss} [{Level}]  (at {Caller}) {Message} {Exception}{NewLine}")
-                .WriteTo.Console(
-                    outputTemplate:
-                    "{Timestamp:HH:mm:ss} [{Level}]  (at {Caller}) {Message} {Exception}{NewLine}")
-                .CreateLogger();
-    }
-
-    
-}
