+++ InternProject.CsvFileConverter/Program.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Autofac;
+using InternProject.CsvFileConverter.Library;
 using Serilog;
 
 namespace CsvFileConverter
